Verificação de IDs DELETE
Deletar ID agora verifica se ele realmente existe ou não

+++ Crud/Crud/src/main/java/AgendaMG/Crud/service/EventoService.java
@@ -38,6 +38,10 @@ public class EventoService {
     }
 
     public void deletarEvento(int id) {
+        if (!eventoRepository.existsById(id)) {
+            throw new RuntimeException("Evento não encontrado");
+        }
+        
         eventoRepository.deleteById(id);
     }
